requirements.txt
torch~=2.1.2
matplotlib~=3.8.2
numpy~=1.26.2
Pillow~=10.1.0
tqdm~=4.66.1
torchvision~=0.16.2
scikit-image~=0.22.0